# Move `Configuration` to `ruff_workspace` crate