--- a/crates/flake8_to_ruff/src/black.rs
+++ b/crates/flake8_to_ruff/src/black.rs
@@ -1,13 +1,12 @@
 //! Extract Black configuration settings from a pyproject.toml.
 
+use ruff::settings::types::PythonVersion;
 use serde::{Deserialize, Serialize};
 
-use crate::settings::types::PythonVersion;
-
 #[derive(Debug, PartialEq, Eq, Serialize, Deserialize, Default)]
-pub struct Black {
+pub(crate) struct Black {
     #[serde(alias = "line-length", alias = "line_length")]
-    pub line_length: Option<usize>,
+    pub(crate) line_length: Option<usize>,
     #[serde(alias = "target-version", alias = "target_version")]
-    pub target_version: Option<Vec<PythonVersion>>,
+    pub(crate) target_version: Option<Vec<PythonVersion>>,
 }