--- a/crates/flake8_to_ruff/src/isort.rs
+++ b/crates/flake8_to_ruff/src/isort.rs
@@ -4,7 +4,7 @@ use serde::{Deserialize, Serialize};
 
 /// The [isort configuration](https://pycqa.github.io/isort/docs/configuration/config_files.html).
 #[derive(Debug, PartialEq, Eq, Serialize, Deserialize, Default)]
-pub struct Isort {
+pub(crate) struct Isort {
     #[serde(alias = "src-paths", alias = "src_paths")]
-    pub src_paths: Option<Vec<String>>,
+    pub(crate) src_paths: Option<Vec<String>>,
 }