--- a/crates/flake8_to_ruff/src/parser.rs
+++ b/crates/flake8_to_ruff/src/parser.rs
@@ -3,11 +3,9 @@ use std::str::FromStr;
 use anyhow::{bail, Result};
 use once_cell::sync::Lazy;
 use regex::Regex;
+use ruff::settings::types::PatternPrefixPair;
+use ruff::{warn_user, RuleSelector};
 use rustc_hash::FxHashMap;
-
-use crate::rule_selector::RuleSelector;
-use crate::settings::types::PatternPrefixPair;
-use crate::warn_user;
 
 static COMMA_SEPARATED_LIST_RE: Lazy<Regex> = Lazy::new(|| Regex::new(r"[,\s]").unwrap());
 
@@ -194,11 +192,11 @@ pub(crate) fn collect_per_file_ignores(
 #[cfg(test)]
 mod tests {
     use anyhow::Result;
+    use ruff::RuleSelector;
 
-    use crate::codes;
-    use crate::registry::Linter;
-    use crate::rule_selector::RuleSelector;
-    use crate::settings::types::PatternPrefixPair;
+    use ruff::codes;
+    use ruff::registry::Linter;
+    use ruff::settings::types::PatternPrefixPair;
 
     use super::{parse_files_to_codes_mapping, parse_prefix_codes, parse_strings};