--- a/crates/flake8_to_ruff/src/pep621.rs
+++ b/crates/flake8_to_ruff/src/pep621.rs
@@ -4,7 +4,7 @@ use pep440_rs::VersionSpecifiers;
 use serde::{Deserialize, Serialize};
 
 #[derive(Debug, PartialEq, Eq, Serialize, Deserialize, Default)]
-pub struct Project {
+pub(crate) struct Project {
     #[serde(alias = "requires-python", alias = "requires_python")]
-    pub requires_python: Option<VersionSpecifiers>,
+    pub(crate) requires_python: Option<VersionSpecifiers>,
 }