--- a/crates/flake8_to_ruff/src/pyproject.rs
+++ b/crates/flake8_to_ruff/src/pyproject.rs
@@ -8,18 +8,18 @@ use super::isort::Isort;
 use super::pep621::Project;
 
 #[derive(Debug, PartialEq, Eq, Serialize, Deserialize)]
-pub struct Tools {
-    pub black: Option<Black>,
-    pub isort: Option<Isort>,
+pub(crate) struct Tools {
+    pub(crate) black: Option<Black>,
+    pub(crate) isort: Option<Isort>,
 }
 
 #[derive(Debug, PartialEq, Eq, Serialize, Deserialize)]
-pub struct Pyproject {
-    pub tool: Option<Tools>,
-    pub project: Option<Project>,
+pub(crate) struct Pyproject {
+    pub(crate) tool: Option<Tools>,
+    pub(crate) project: Option<Project>,
 }
 
-pub fn parse<P: AsRef<Path>>(path: P) -> Result<Pyproject> {
+pub(crate) fn parse<P: AsRef<Path>>(path: P) -> Result<Pyproject> {
     let contents = std::fs::read_to_string(path)?;
     let pyproject = toml::from_str::<Pyproject>(&contents)?;
     Ok(pyproject)