--- a/crates/ruff/src/lib.rs
+++ b/crates/ruff/src/lib.rs
@@ -12,13 +12,12 @@ pub const VERSION: &str = env!("CARGO_PKG_VERSION");
 
 mod autofix;
 mod checkers;
-mod codes;
+pub mod codes;
 mod comments;
 mod cst;
 pub mod directives;
 mod doc_lines;
 mod docstrings;
-pub mod flake8_to_ruff;
 pub mod fs;
 mod importer;
 pub mod jupyter;
@@ -33,7 +32,7 @@ pub mod pyproject_toml;
 pub mod registry;
 mod renamer;
 mod rule_redirects;
-mod rule_selector;
+pub mod rule_selector;
 pub mod rules;
 pub mod settings;
 pub mod source_kind;
@@ -41,3 +40,5 @@ pub mod upstream_categories;
 
 #[cfg(any(test, fuzzing))]
 pub mod test;
+
+pub const RUFF_PKG_VERSION: &str = env!("CARGO_PKG_VERSION");