--- a/crates/ruff/src/packaging.rs
+++ b/crates/ruff/src/packaging.rs
@@ -2,10 +2,6 @@
 
 use std::path::{Path, PathBuf};
 
-use rustc_hash::FxHashMap;
-
-use crate::resolver::{PyprojectConfig, Resolver};
-
 // If we have a Python package layout like:
 // - root/
 //   - foo/
@@ -49,68 +45,6 @@ pub fn detect_package_root<'a>(
         current = Some(parent);
     }
     current
-}
-
-/// A wrapper around `is_package` to cache filesystem lookups.
-fn is_package_with_cache<'a>(
-    path: &'a Path,
-    namespace_packages: &'a [PathBuf],
-    package_cache: &mut FxHashMap<&'a Path, bool>,
-) -> bool {
-    *package_cache
-        .entry(path)
-        .or_insert_with(|| is_package(path, namespace_packages))
-}
-
-/// A wrapper around `detect_package_root` to cache filesystem lookups.
-fn detect_package_root_with_cache<'a>(
-    path: &'a Path,
-    namespace_packages: &'a [PathBuf],
-    package_cache: &mut FxHashMap<&'a Path, bool>,
-) -> Option<&'a Path> {
-    let mut current = None;
-    for parent in path.ancestors() {
-        if !is_package_with_cache(parent, namespace_packages, package_cache) {
-            return current;
-        }
-        current = Some(parent);
-    }
-    current
-}
-
-/// Return a mapping from Python package to its package root.
-pub fn detect_package_roots<'a>(
-    files: &[&'a Path],
-    resolver: &'a Resolver,
-    pyproject_config: &'a PyprojectConfig,
-) -> FxHashMap<&'a Path, Option<&'a Path>> {
-    // Pre-populate the module cache, since the list of files could (but isn't
-    // required to) contain some `__init__.py` files.
-    let mut package_cache: FxHashMap<&Path, bool> = FxHashMap::default();
-    for file in files {
-        if file.ends_with("__init__.py") {
-            if let Some(parent) = file.parent() {
-                package_cache.insert(parent, true);
-            }
-        }
-    }
-
-    // Search for the package root for each file.
-    let mut package_roots: FxHashMap<&Path, Option<&Path>> = FxHashMap::default();
-    for file in files {
-        let namespace_packages = &resolver.resolve(file, pyproject_config).namespace_packages;
-        if let Some(package) = file.parent() {
-            if package_roots.contains_key(package) {
-                continue;
-            }
-            package_roots.insert(
-                package,
-                detect_package_root_with_cache(package, namespace_packages, &mut package_cache),
-            );
-        }
-    }
-
-    package_roots
 }
 
 #[cfg(test)]