--- a/crates/ruff/src/rules/flake8_tidy_imports/mod.rs
+++ b/crates/ruff/src/rules/flake8_tidy_imports/mod.rs
@@ -1,6 +1,5 @@
 //! Rules from [flake8-tidy-imports](https://pypi.org/project/flake8-tidy-imports/).
 pub(crate) mod matchers;
-pub mod options;
 pub(crate) mod rules;
 pub mod settings;