… Cargo.lock
… crates/flake8_to_ruff/Cargo.toml
… crates/ruff/Cargo.toml
… crates/ruff/src/flake8_to_ruff/external_config.rs
… crates/ruff/src/flake8_to_ruff/mod.rs
… crates/ruff/src/rules/flake8_annotations/settings.rs
… crates/ruff/src/rules/flake8_bugbear/settings.rs
… crates/ruff/src/rules/flake8_builtins/settings.rs
… crates/ruff/src/rules/flake8_comprehensions/settings.rs
… crates/ruff/src/rules/flake8_copyright/settings.rs
… crates/ruff/src/rules/flake8_errmsg/settings.rs
… crates/ruff/src/rules/flake8_gettext/settings.rs
… crates/ruff/src/rules/flake8_implicit_str_concat/settings.rs
… crates/ruff/src/rules/flake8_import_conventions/mod.rs
… crates/ruff/src/rules/flake8_import_conventions/settings.rs
… crates/ruff/src/rules/flake8_pytest_style/settings.rs
… crates/ruff/src/rules/flake8_quotes/settings.rs
… crates/ruff/src/rules/flake8_self/settings.rs
… crates/ruff/src/rules/flake8_tidy_imports/options.rs
… crates/ruff/src/rules/flake8_type_checking/settings.rs
… crates/ruff/src/rules/flake8_unused_arguments/settings.rs
… crates/ruff/src/rules/isort/categorize.rs
… crates/ruff/src/rules/isort/settings.rs
… crates/ruff/src/rules/mccabe/settings.rs
… crates/ruff/src/rules/pep8_naming/settings.rs
… crates/ruff/src/rules/pycodestyle/settings.rs
… crates/ruff/src/rules/pydocstyle/settings.rs
… crates/ruff/src/rules/pyflakes/settings.rs
… crates/ruff/src/rules/pylint/settings.rs
… crates/ruff/src/rules/pyupgrade/settings.rs