<!--
Thank you for contributing to Ruff! To help us out with reviewing, please consider the following:

- Does this pull request include a summary of the change? (See below.)
- Does this pull request include a descriptive title?
- Does this pull request include references to any relevant issues?
-->

## Summary

This PR introduces a new `ruff_workspace` crate and moves `Configuration` and the project file discovery to the newly created crate. 

The motivation for this change is to 

a) Speed up the ruff library compile time to enable faster development cycles
b) Prepare for the formatter where it becomes necessary to add formatter specific configurations to `Configuration`. This wouldn't be possible today without adding `ruff_python_formatter` as a dependency to `ruff`, which we want to avoid (even slower compile times). 

Some consequences of the refactor:

* It was possible to move the whole `flake8` converter module from ruff to the flake8 crate.
* All plugin options are no longer defined alongside their rule code but had to be moved to `ruff_workspace`. I think that's how it is intended.


## Results

This reduces the release llvm lines from 1801126 -> 1451732 (~20% reduction) and building the ruff library in release mode drops from 43s to 33s on my machine, which is a lot! 

It also reduces the ruff binary release build time from 156 to 137s, mainly because `ruff_workspace` can compile in parallel to `ruff` (at least part of the time)

## Test Plan

`cargo build` and `cargo test`

<!-- How was it tested? -->