@MichaReiser 2023-08-27
Current dependencies on/for this PR:
* main
  * **PR #6920** <a href="https://app.graphite.dev/github/pr/astral-sh/ruff/6920" target="_blank"><img src="https://static.graphite.dev/graphite-32x32.svg" alt="Graphite" width="10px" height="10px"/></a>  👈

This comment was auto-generated by [Graphite](https://app.graphite.dev/github/pr/astral-sh/ruff/6920?utm_source=stack-comment).